Reduce unnecessary render() calls by ensuring ref equality with style objects.
Adds a style objects cache to the Grid Component, which is used by the
defaultCellRangeRenderer to recover previously generated styles rather
than generate new objects when appropriate. This can significantly
reduce the number of calls to the cell component's render() function
when shallowCompare is used, because the style object will maintain
referential equality.